Refactor components and update TypeScript configuration for improved structure and compatibility
References:
#6
