+++ front-end/nhl-app/package.json
@@ -21,7 +21,7 @@
   },
   "devDependencies": {
     "@types/node": "latest",
-    "@types/react": "latest",
+    "@types/react": "^19.0.10",
     "@types/react-dom": "latest",
     "eslint": "latest",
     "eslint-config-next": "^14.0.0",
