+++ front-end/nhl-app/src/app/components/TeamLogoWithLabel/TeamLogoWithLabel.tsx
